Refactor file paths for database resources

+++ proftrack/src/main/java/ma/ac/usms/ensak/presentation/controller/DocumentController.java
@@ -1,7 +1,7 @@
 package ma.ac.usms.ensak.presentation.controller;
 
 import ma.ac.usms.ensak.metier.management.DocumentManager;
-
+import ma.ac.usms.ensak.exception.*;
 
 public class DocumentController {
     private static DocumentManager documentManager = new DocumentManager();
@@ -16,11 +16,15 @@ public class DocumentController {
         return documentBox;
     }
 
-    
-
     public void deleteDocument() {
         documentBox.getDocumentBox().getDeleteButton().setOnAction(e -> {
-            documentManager.deleteDocument(null);
+            if (ConfirmationDialog.showConfirmationDialog("Are you sure you want to delete this document?")) {
+                documentManager.deleteDocument(null);
+                AlertHandler.showSuccessAlert("Document deleted successfully");
+            } else {
+                AlertHandler.showFailureAlert("Document deletion canceled");
+
+            }
         });
     }
